requirements.txt
appdirs==1.4.4
liftover==1.3.1
PySimpleGUI==4.70.1
PyYAML==6.0.2
urllib3==2.3.0
